+++ texar/agents/dqn_agent.py
@@ -227,8 +227,7 @@ class DQNAgent(EpisodicAgentBase):
 
             - If **"copy"**, the target network is assigned with the parameter \
             of Q-net every :attr:`"update_period"` steps.
-
-            - If **"tau"**, target will be updated by assigning as
+            - If **"tau"**, target will be updated by assigning as \
             ``` (1 - 1/update_period) * target + 1/update_period * qnet ```
 
         "update_period": int
@@ -292,7 +291,7 @@ class DQNAgent(EpisodicAgentBase):
         return self._target(inputs=state_inputs, **self._qnet_caller_kwargs)
 
     def _get_td_error(self, qnet_qvalues, actions, y):
-        return y - tf.reduce_sum(qnet_qvalues * tf.to_float(actions), axis=1)
+        return y - tf.reduce_sum(qnet_qvalues * tf.cast(actions, tf.float), axis=1)
 
     def _get_train_op(self):
         train_op = opt.get_train_op(