+++ texar/models/model_base.py
@@ -19,7 +19,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from texar import HParams
+from texar.hyperparams import HParams
 
 # pylint: disable=too-many-arguments